requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# User Agent Parsing
user-agents==2.2.0

# HTTP Requests for Webhooks
requests==2.31.0

# Additional Flask Extensions
Flask-CORS==4.0.0

# Utility Dependencies
click==8.1.7
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Security and Cryptography
cryptography==41.0.7

# JSON Processing
simplejson==3.19.2

# Date and Time Utilities
python-dateutil==2.8.2

# Environment Variable Management
python-dotenv==1.0.0
